+++ components/render/index.js
@@ -1,0 +1,4 @@
+// Barrel for the render layer.
+export { default as Texture } from "./texture.js";
+export { default as SpriteSheet, Animation, Animator } from "./spriteSheet.js";
+export { getProgramInfo, PROGRAM_COLOR, PROGRAM_TEXTURE } from "./shaders.js";